Add NER class
