+++ pythainlp/tag/__init__.py
@@ -11,9 +11,12 @@ __all__ = [
     "pos_tag",
     "pos_tag_sents",
     "tag_provinces",
+    "chunk_parse",
+    "NER",
 ]
 
 from pythainlp.tag.locations import tag_provinces
 from pythainlp.tag.pos_tag import pos_tag, pos_tag_sents
 from pythainlp.tag._tag_perceptron import PerceptronTagger
 from pythainlp.tag.chunk import chunk_parse
+from pythainlp.tag.named_entity import NER
